cleanup histogram code a little

+++ src/types.odin
@@ -86,7 +86,7 @@ Stats :: struct {
 	min_time: f64,
 	max_time: f64,
 	count: u32,
-	hist: [100] u32,
+	hist: [100]f64,
 }
 Range :: struct {
 	pid: int,
